refactor: 💡 Requestをparserのインナークラスに変更
